requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
autograd==1.7.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.20.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.14.4
DeepXDE==1.12.2
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.2.0
inspect-it==0.3.2
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.11.9
lion-pytorch==0.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.555.43
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.13.1
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phi==0.6.7
phi-torch==0.0.4
phiflow==3.0.0
phiml==1.7.0.post1
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaml==24.9.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyDOE==0.3.8
pyDOE2==1.3.0
pyfiglet==1.0.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.14.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensordiffeq==0.2.0
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.2.2
torchdiffeq==0.2.2
torchmetrics==1.6.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wandb==0.18.7
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.11
wrapt==1.17.0
xxhash==3.4.1
yarl==1.9.4
yasm==0.1.0a1
zipp==3.19.2
